refactor(LadderResultController): Controller 리팩토링
- Ladder의 결과인 LadderResult 만드는 책임 부여
- LadderGame, Participants 폴더 생성
- Participants에서 String -> Name 변환 method 이름 불문명한 것 같아 변경
- Participant의 크기를 주는 method size()로 직관적으로 변경 (try...)
- 생성자에서 확인 가능하도록 validateMethod 바로 아래에 위치시키도록 이동
