+++ src/main/java/ladder/domain/participant/Participants.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.domain.participant;
 
 import java.util.Collections;
 import java.util.HashSet;
@@ -8,19 +8,23 @@ import java.util.Set;
 public class Participants {
 
     private static final int MAXIMUM_RECEPTION_COUNT = 10;
-
     private final List<Name> names;
 
     public Participants(List<String> inputNames) {
         validate(inputNames);
-        this.names = inviteParticipants(inputNames);
+        this.names = enrollNames(inputNames);
+    }
+
+    private void validate(List<String> names) {
+        validateParticipantIsNotOverMaximumCount(names);
+        validateHasNotDuplicatedName(names);
     }
 
     public List<Name> getNames() {
         return Collections.unmodifiableList(names);
     }
 
-    public int getParticipantsCount() {
+    public int size() {
         return names.size();
     }
 
@@ -33,24 +37,19 @@ public class Participants {
         return index;
     }
 
-    private List<Name> inviteParticipants(List<String> names) {
+    private List<Name> enrollNames(List<String> names) {
         return names.stream()
                 .map(Name::new)
                 .toList();
     }
 
-    private void validate(List<String> names) {
-        validateIsNotOverMaximumReception(names);
-        validateHasNotDuplicateName(names);
-    }
-
-    private void validateIsNotOverMaximumReception(List<String> names) {
+    private void validateParticipantIsNotOverMaximumCount(List<String> names) {
         if (names.size() > MAXIMUM_RECEPTION_COUNT) {
             throw new IllegalArgumentException("참여자 인원은 최대 " + MAXIMUM_RECEPTION_COUNT + "명까지 가능합니다.");
         }
     }
 
-    private void validateHasNotDuplicateName(List<String> names) {
+    private void validateHasNotDuplicatedName(List<String> names) {
         Set<String> uniqueNames = new HashSet<>(names);
 
         if (names.size() != uniqueNames.size()) {
